LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/securitypolicyviolation-block-cross-origin-image.sub-expected.txt


FAIL Non-redirected cross-origin URLs are not stripped. assert_equals: expected 5 but got 6
